[DEV-1141] Added swagger docu
