+++ src/subdomains/generic/user/models/wallet/dto/kyc-data.dto.ts
@@ -1,4 +1,4 @@
-import { KycWebhookStatus } from '../../../services/webhook/webhook.service';
+import { KycWebhookStatus } from '../../../services/webhook/dto/kyc-webhook.dto';
 
 export class KycDataDto {
   address: string;
