+++ src/subdomains/generic/user/services/webhook/dto/webhook.dto.ts
@@ -1,0 +1,21 @@
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+
+export enum WebhookType {
+  PAYMENT = 'Payment',
+  KYC_CHANGED = 'KycChanged',
+  KYC_FAILED = 'KycFailed',
+}
+
+export class WebhookDto<T> {
+  @ApiProperty()
+  id: string;
+
+  @ApiProperty({ enum: WebhookType })
+  type: WebhookType;
+
+  @ApiProperty()
+  data: T;
+
+  @ApiPropertyOptional()
+  reason: string;
+}
